+++ synphot/tests/test_spectrum.py
@@ -1091,6 +1091,7 @@ class TestWriteSpec:
         shutil.rmtree(self.outdir)
 
 
+@pytest.mark.skipif('not HAS_SCIPY')
 @pytest.mark.skipif('not HAS_SPECUTILS')
 class TestSpecutilsBridge:
     def test_from_spectrum1d_Empirical1D_source(self):