+++ CHANGES.rst
@@ -4,6 +4,8 @@
 - ``SourceSpectrum`` and ``SpectralElement`` now have ``to_spectrum1d`` and
   ``from_spectrum1d`` to write to and read from ``specutils.Spectrum1D``
   object, respectively. [#243]
+- ``Observation`` now accepts ``specutils.Spectrum1D`` as a valid source
+  spectrum input. [#246]
 - Dropped support for Python 3.5 and ``astropy`` 2.x. This version is only
   compatible with Python 3.6 or later and ``astropy`` 3.x or later. [#243]
 - Added support for ``RickerWavelet1D`` model that is the renamed version